Start of album data
