+++ package.json
@@ -49,6 +49,7 @@
     "@types/react": "^17.0.34",
     "@types/react-dom": "^17.0.11",
     "@types/react-helmet": "^6.1.4",
+    "gatsby-plugin-ts-config": "^2.1.3",
     "markdownlint": "^0.24.0",
     "markdownlint-cli": "^0.29.0",
     "prettier": "^2.4.1",
