+++ src/components/header/Header.tsx
@@ -8,8 +8,8 @@ const Header = () => (
             <h1 className={styles.title}>Zoe Aubert</h1>
         </Link>
         <div className={styles.links}>
-            <Link className={styles.link} to="/blog">Blog</Link>
             <Link className={styles.link} to="/photos">Photos</Link>
+            <Link className={styles.link} to="/blog">Blog</Link>
         </div>
     </div>
 )
