+++ src/components/header/header.module.scss
@@ -3,9 +3,9 @@
 $headerHeight: 2rem;
 
 .container {
-  flex-direction: row;
+  flex-direction: column;
   flex-wrap: nowrap;
-  justify-content: space-between;
+  justify-content: center;
   align-items: baseline;
   width: 100%;
   max-width: $max-width;
@@ -13,13 +13,16 @@ $headerHeight: 2rem;
 }
 
 .titleContainer {
+  width: 100%;
+  justify-content: center;
   align-items: center;
 }
 
 .title {
-  font-size: 140%;
+  font-size: 150%;
   margin: 0;
   padding: .5rem 0;
+  text-align: center;
 
   transition: text-decoration .25s;
 
@@ -29,17 +32,20 @@ $headerHeight: 2rem;
 }
 
 .links {
-  height: $headerHeight;
   flex-direction: row;
   flex-wrap: nowrap;
-  justify-content: flex-end;
+  justify-content: space-around;
   align-items: center;
+  width: 100%;
+  height: $headerHeight;
 }
 
 .link {
+  flex: 1;
   font-weight: normal;
   margin: 0 0 0 .5rem;
   transition: text-decoration .25s;
+  text-align: center;
 
   &:hover {
     text-decoration: underline;
@@ -48,6 +54,44 @@ $headerHeight: 2rem;
 
 @media only screen and (min-width: $collapsable-width) {
   .container {
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+  }
+
+  .titleContainer {
+    width: auto;
+  }
+
+  .title {
+    font-size: 150%;
+    margin: 0;
     padding: .5rem 0;
+    text-align: left;
+
+    transition: text-decoration .25s;
+
+    &:hover {
+      text-decoration: underline;
+    }
+  }
+
+  .links {
+    height: $headerHeight;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: flex-end;
+    align-items: center;
+  }
+
+  .link {
+    flex: 0;
+    font-weight: normal;
+    margin: 0 0 0 .5rem;
+    transition: text-decoration .25s;
+
+    &:hover {
+      text-decoration: underline;
+    }
   }
 }
