+++ src/components/page/page.module.scss
@@ -13,6 +13,5 @@
   align-items: center;
   width: 100%;
   max-width: $max-width;
-  //background-color: blue;
+  padding: 0 1rem;
 }
-
