+++ src/templates/AlbumPage.tsx
@@ -1,0 +1,7 @@
+
+
+const AlbumPage = () => {
+    return null
+}
+
+export default AlbumPage
